requirements.txt
yfinance
pandas
numpy
matplotlib
scikit-learn
joblib
python-dateutil